Added a check for meta key existance for "Remove Thumbnail" button action. This prevents it from displaying "AJAX Error" message if user is trying to delete a thumbnail when it doesn't exists.

+++ simple-download-monitor/main.php
@@ -231,9 +231,9 @@ class simpleDownloadManager {
             <?php
             // Localize langauge strings used in js file
             $sdmTranslations = array(
-                'select_file' =>  __('Select File', 'simple-download-monitor'),
-                'select_thumbnail' =>  __('Select Thumbnail', 'simple-download-monitor'),
-                'insert' =>  __('Insert', 'simple-download-monitor'),
+                'select_file' => __('Select File', 'simple-download-monitor'),
+                'select_thumbnail' => __('Select Thumbnail', 'simple-download-monitor'),
+                'insert' => __('Insert', 'simple-download-monitor'),
                 'image_removed' => __('Image Successfully Removed', 'simple-download-monitor'),
                 'ajax_error' => __('Error with AJAX', 'simple-download-monitor')
             );
@@ -708,9 +708,15 @@ function sdm_remove_thumbnail_image_ajax_call() {
 
 //Go ahead with the thumbnail removal
     $post_id = $_POST['post_id_del'];
-    $success = delete_post_meta($post_id, 'sdm_upload_thumbnail');
-    if ($success) {
-        $response = json_encode(array('success' => true));
+    $key_exists = metadata_exists('post', $post_id, 'sdm_upload_thumbnail');
+    if ($key_exists) {
+        $success = delete_post_meta($post_id, 'sdm_upload_thumbnail');
+        if ($success) {
+            $response = json_encode(array('success' => true));
+        }
+    } else {
+        // in order for frontend script to not display "Ajax error", let's return some data
+        $response = json_encode(array('not_exists' => true));
     }
 
     header('Content-Type: application/json');
